Moving common tests into test_common dir
