+++ test/test_common/test_request.py
@@ -104,7 +104,7 @@ def test_add_extracted(service_request):
 
 
 def test_add_image(service_request):
-    image_path = "./test/b32969aa664e3905c20f865cdd7b921f922678f5c3850c78e4c803fbc1757a8e"
+    image_path = "./test/test_common/b32969aa664e3905c20f865cdd7b921f922678f5c3850c78e4c803fbc1757a8e"
 
     # Basic
     assert service_request.add_image(image_path, "image_name", "description of image") == {
@@ -192,7 +192,7 @@ def test_add_image(service_request):
     service_request.task.supplementary.clear()
 
     # Classification, OCR heuristic, OCR_IO, image with password
-    image_path = "./test/4031ed8786439eee24b87f84901e38038a76b8c55e9d87dd5a7d88df2806c1cf"
+    image_path = "./test/test_common/4031ed8786439eee24b87f84901e38038a76b8c55e9d87dd5a7d88df2806c1cf"
     _, path = tempfile.mkstemp()
     ocr_io = open(path, "w")
     data = service_request.add_image(image_path, "image_name", "description of image", "TLP:A", ocr_heuristic_id, ocr_io)
