+++ test/test_common/test_result.py
@@ -581,7 +581,7 @@ def test_imagesectionbody_init(service_request):
 
 def test_imagesectionbody_add_image(service_request):
     isb = ImageSectionBody(service_request)
-    image_path = "./test/b32969aa664e3905c20f865cdd7b921f922678f5c3850c78e4c803fbc1757a8e"
+    image_path = "./test/test_common/b32969aa664e3905c20f865cdd7b921f922678f5c3850c78e4c803fbc1757a8e"
 
     # Basic
     assert isb.add_image(image_path, "image_name", "description of image") is None
@@ -1207,7 +1207,7 @@ def test_resultimagesection_init(service_request):
 def test_resultimagesection_add_image(service_request):
     ris = ResultImageSection(service_request, "title_text_as_str")
 
-    image_path = "./test/b32969aa664e3905c20f865cdd7b921f922678f5c3850c78e4c803fbc1757a8e"
+    image_path = "./test/test_common/b32969aa664e3905c20f865cdd7b921f922678f5c3850c78e4c803fbc1757a8e"
 
     # Basic
     assert ris.add_image(image_path, "image_name", "description of image") is None
